Move worksRouteToApiUrl into api.ts
