+++ catalogue/webapp/pages/works.tsx
@@ -5,7 +5,7 @@ import { grid, classNames } from '@weco/common/utils/classnames';
 import convertUrlToString from '@weco/common/utils/convert-url-to-string';
 import CataloguePageLayout from '../components/CataloguePageLayout/CataloguePageLayout';
 import Paginator from '@weco/common/views/components/Paginator/Paginator';
-import { worksRouteToApiUrl } from '@weco/common/services/catalogue/ts_api';
+import { worksRouteToApiUrl } from '@weco/common/services/catalogue/api';
 import Space from '@weco/common/views/components/styled/Space';
 import { getWorks } from '../services/catalogue/works';
 import cookies from 'next-cookies';
