+++ common/services/catalogue/api.ts
@@ -1,4 +1,6 @@
-import { WorksRouteProps } from './routes';
+import { quoteVal } from '../../utils/csv';
+import { ImagesProps } from '../../views/components/ImagesLink/ImagesLink';
+import { WorksProps } from '../../views/components/WorksLink/WorksLink';
 
 function toIsoDateString(s: string | undefined): string | undefined {
   if (s) {
@@ -11,32 +13,39 @@ function toIsoDateString(s: string | undefined): string | undefined {
   return undefined;
 }
 
+type ItemsLocationsAccessConditionsStatus =
+  | 'open'
+  | 'open-with-advisory'
+  | 'restricted'
+  | 'closed'
+  | 'licensed-resources'
+  | 'unavailable'
+  | 'permission-required'
+  | '!open'
+  | '!open-with-advisory'
+  | '!restricted'
+  | '!closed'
+  | '!licensed-resources'
+  | '!unavailable'
+  | '!permission-required';
+
 export type CatalogueWorksApiProps = {
   query?: string;
   page?: number;
   workType?: string[];
   'items.locations.locationType'?: string[];
-  'items.locations.accessConditions.status'?: (
-    | 'open'
-    | 'open-with-advisory'
-    | 'restricted'
-    | 'closed'
-    | 'licensed-resources'
-    | 'unavailable'
-    | 'permission-required'
-    | '!open'
-    | '!open-with-advisory'
-    | '!restricted'
-    | '!closed'
-    | '!licensed-resources'
-    | '!unavailable'
-    | '!permission-required'
-  )[];
+  'items.locations.accessConditions.status'?: ItemsLocationsAccessConditionsStatus;
   availabilities?: string[];
   sort?: string;
   sortOrder?: string;
+  'partOf.title'?: string;
   'production.dates.from'?: string;
   'production.dates.to'?: string;
+  'genres.label'?: string[];
+  'subjects.label'?: string[];
+  subjects?: string[];
+  'contributors.agent.label'?: string[];
+  languages?: string[];
   _queryType?: string;
   aggregations?: string[];
 };
@@ -56,21 +65,29 @@ export type CatalogueConceptsApiProps = {
 };
 
 export function worksRouteToApiUrl(
-  worksRouteProps: WorksRouteProps,
-  overrides: CatalogueWorksApiProps
+  worksProps: WorksProps,
+  overrides: Partial<CatalogueWorksApiProps>
 ): CatalogueWorksApiProps {
   return {
-    query: worksRouteProps.query,
-    page: worksRouteProps.page,
-    workType: worksRouteProps.workType,
-    'items.locations.locationType': worksRouteProps.itemsLocationsLocationType,
-    availabilities: worksRouteProps.availabilities,
-    sort: worksRouteProps.sort,
-    sortOrder: worksRouteProps.sortOrder,
-    'production.dates.from': toIsoDateString(
-      worksRouteProps.productionDatesFrom
-    ),
-    'production.dates.to': toIsoDateString(worksRouteProps.productionDatesTo),
+    query: worksProps.query,
+    page: worksProps.page,
+    workType: worksProps.workType,
+    'items.locations.locationType': worksProps['items.locations.locationType'],
+    availabilities: worksProps.availabilities,
+    sort: worksProps.sort,
+    sortOrder: worksProps.sortOrder,
+    'partOf.title': worksProps['partOf.title'],
+    'production.dates.from': worksProps['production.dates.from']
+      ? toIsoDateString(worksProps['production.dates.from'])
+      : undefined,
+    'production.dates.to': worksProps['production.dates.to']
+      ? toIsoDateString(worksProps['production.dates.to'])
+      : undefined,
+    languages: worksProps.languages,
+    'genres.label': worksProps['genres.label'].map(quoteVal),
+    'subjects.label': worksProps['subjects.label'].map(quoteVal),
+    'contributors.agent.label':
+      worksProps['contributors.agent.label'].map(quoteVal),
     ...overrides,
   };
 }
